Point 4.1 done, current sale done updating, also changed Date with Calendar Class.

+++ Lab4/src/main/java/UPF_POO20_G101_20/Lab4/users/Seller.java
@@ -17,12 +17,12 @@ public class Seller extends User{
         accountNumber = a;
     }
 
-    public void sell(Item i, int number) {
+    public void sell(Item i, double number) {
         if (deposit(i.getPrice()) && hasItem(i)) {
 			double x = 0.0;
         	if (i instanceof UnitItem) {
         		UnitItem item = (UnitItem)i;
-        		x = item.sell(number);
+        		x = item.sell((int)number);
         	}
         	else if (i instanceof WeightedItem) {
         		WeightedItem item = (WeightedItem)i;
